backlog: rename storage -> memstorage package
